Type Promotion Rule byte,short,char promoted to int
if one operand is long, whole expression promoted to long, if one operand is float, entire expression is promoted to float, if any operand is double, result is double.  In first term f*b, b is promoted to a float and result is float.  In second term, c is promoted to int and result is int. In third term, s is promoted to double, result is double. Finally, when float, int and double are added,  float+int results in float,  float-double results in double, double+double gives double.